pyproject.toml
[project]
name = "airline-checkin-agent"
version = "0.1.0"
description = "AI agent for automated airline check-in using Strands SDK"
requires-python = ">=3.11"
dependencies = [
    "strands-agents==1.20.0",
    "strands-agents-tools==0.2.18",
    "playwright>=1.40.0",
    "boto3>=1.34.0",
    "numpy>=1.24.0",
    "pydantic>=2.0.0",
    "nest-asyncio>=1.5.0",
    "pillow>=10.0.0",
    "PyYAML>=6.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "aws-opentelemetry-distro>=0.10.1",
    "bedrock-agentcore>=1.1.1",
    "strands-agents[otel]",
    "cryptography>=41.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.1.0",
]
